Implementing new tests
Implementing unit tests from the methods in the pets_repository, now using mock instead of changing the actual database. Checking if the connection is calling PEtstable, if the all function is beeing called, and if the filter function is bot beeing called. Also asserting if the value is matching the required data